Initial support SLQSSDK

+++ source/libmodem/modem_db.h
@@ -17,19 +17,19 @@ typedef void *(*pthread_func_t)(void*);
 
 /*------------------------------------------------------------------------*/
 
-typedef char* (*get_imei_func_t)(modem_t* modem, char* imei, int len);
+typedef char* (*get_imei_func_t)(modem_t* modem, char* imei, size_t len);
 
 typedef int (*get_signal_quality_func_t)(modem_t* modem, modem_signal_quality_t* sq);
 
 typedef time_t (*get_network_time_func_t)(modem_t* modem);
 
-typedef char* (*get_imsi_func_t)(modem_t* modem, char* imsi, int len);
+typedef char* (*get_imsi_func_t)(modem_t* modem, char* imsi, size_t len);
 
-typedef char* (*get_operator_name_func_t)(modem_t* modem, char *oper, int len);
+typedef char* (*get_operator_name_func_t)(modem_t* modem, char *oper, size_t len);
 
 typedef modem_network_reg_t (*network_registration_func_t)(modem_t* modem);
 
-typedef char* (*get_network_type_func_t)(modem_t* modem, char *network, int len);
+typedef char* (*get_network_type_func_t)(modem_t* modem, char *network, size_t len);
 
 typedef int (*change_pin_func_t)(modem_t* modem, const char* old_pin, const char* new_pin);
